Got the injestion working (bad port #). refactored Poller a bit
