+++ ingestor/src/main/java/app/Application.java
@@ -9,8 +9,7 @@ public class Application {
     Poller poller = new Poller(0,
       10000,
       "https://crest-tq.eveonline.com/market/prices/",
-      new WriteToRethinkCallback("eve1", "marketstat1", 9090, "admin"));
+      new WriteToRethinkCallback("eve", "market-prices", 28015, "admin"));
     poller.poll();
-    System.out.println("app running");
   }
 }
