+++ ingestor/src/main/java/app/callback/WriteToRethinkCallback.java
@@ -8,6 +8,9 @@ import org.json.simple.parser.ParseException;
 
 import java.io.IOException;
 
+/**
+ * TODO
+ */
 public class WriteToRethinkCallback implements Callback {
 
   RethinkClient rc;
@@ -16,24 +19,22 @@ public class WriteToRethinkCallback implements Callback {
   JSONParser jp = new JSONParser();
 
   public WriteToRethinkCallback(String dbName, String tableName, Integer port, String username) {
-    this.rc = new RethinkClient(port, username); // TODO figure out user name
+    this.rc = new RethinkClient(port, username);
     this.dbName = dbName;
     this.tableName = tableName;
-
     this.rc.createDatabase(this.dbName);
     this.rc.createTable(this.dbName, this.tableName);
   }
 
   public void call(Response response) {
     try {
-      this.rc.insertData(this.dbName, this.tableName, response.body().string());
-      // this.rc.insertData(this.dbName, this.tableName, jp.parse(response.body().string()));
+      this.rc.insertData(this.dbName, this.tableName, jp.parse(response.body().string()));
     } catch (IOException e) {
       System.err.println("Error inserting data into database");
       e.printStackTrace();
-    // } catch (ParseException pe) {
-    //   System.err.println("Error parsing response");
-    //   pe.printStackTrace();
+    } catch (ParseException pe) {
+      System.err.println("Error parsing response");
+      pe.printStackTrace();
     }
   }
 }
